Refactor User Profile details
