+++ ts/features/profile/navigation/navigator.ts
@@ -6,11 +6,13 @@ import {
 } from "react-navigation-stack/src/types";
 import ProfileDeletionWarningScreen from "../screens/ProfileDeletionWarningScreen";
 import { newProfileScreenEnabled } from "../../../config";
+import ProfileDeletionSummaryScreen from "../screens/ProfileDeletionSummaryScreen";
 import PROFILE_DELETION_ROUTES from "./routes";
 
 export const ProfileDeletionNavigator = createStackNavigator(
   {
-    [PROFILE_DELETION_ROUTES.START]: { screen: ProfileDeletionWarningScreen }
+    [PROFILE_DELETION_ROUTES.START]: { screen: ProfileDeletionWarningScreen },
+    [PROFILE_DELETION_ROUTES.CONFIRM]: { screen: ProfileDeletionSummaryScreen }
   },
   {
     headerMode: "none",
