+++ ts/features/profile/screens/ProfileDeletionSummaryScreen.tsx
@@ -1,30 +1,22 @@
 import React from "react";
 import { Dispatch } from "redux";
 import { connect } from "react-redux";
-import { Content } from "native-base";
 import { SafeAreaView } from "react-native";
 import { NavigationStackScreenProps } from "react-navigation-stack";
 import { GlobalState } from "../../../store/reducers/types";
 import { emptyContextualHelp } from "../../../utils/emptyContextualHelp";
 import BaseScreenComponent from "../../../components/screens/BaseScreenComponent";
 import { IOStyles } from "../../../components/core/variables/IOStyles";
-import { H1 } from "../../../components/core/typography/H1";
 import I18n from "../../../i18n";
-import { H4 } from "../../../components/core/typography/H4";
 import FooterWithButtons from "../../../components/ui/FooterWithButtons";
 import { userProfileDeletionCancel } from "../actions";
-import PROFILE_DELETION_ROUTES from "../navigation/routes";
-import { useNavigationContext } from "../../../utils/hooks/useOnFocus";
+import UserProfileDetails from "../components/UserProfileDetails";
 
 type Props = ReturnType<typeof mapDispatchToProps> &
   ReturnType<typeof mapStateToProps> &
   NavigationStackScreenProps;
 
-const ProfileDeletionWarningScreen = (props: Props): React.ReactElement => {
-  const navigationContext = useNavigationContext();
-  const confirmDeletion = () =>
-    navigationContext.navigate(PROFILE_DELETION_ROUTES.CONFIRM);
-
+const ProfileDeletionSummaryScreen = (props: Props): React.ReactElement => {
   const cancelButtonProps = {
     block: true,
     light: true,
@@ -36,19 +28,14 @@ const ProfileDeletionWarningScreen = (props: Props): React.ReactElement => {
   const continueButtonProps = {
     block: true,
     primary: true,
-    onPress: confirmDeletion,
+    onPress: undefined,
     title: I18n.t("global.buttons.continue")
   };
 
   return (
     <BaseScreenComponent goBack={true} contextualHelp={emptyContextualHelp}>
       <SafeAreaView style={IOStyles.flex}>
-        <Content>
-          <H1>{I18n.t("profile.main.privacy.removeAccount.title")}</H1>
-          <H4 weight="Regular">
-            {I18n.t("profile.main.privacy.removeAccount.info.body")}
-          </H4>
-        </Content>
+        <UserProfileDetails></UserProfileDetails>
         <FooterWithButtons
           type="TwoButtonsInlineThird"
           leftButton={cancelButtonProps}
@@ -68,4 +55,4 @@ const mapStateToProps = (state: GlobalState) => ({});
 export default connect(
   mapStateToProps,
   mapDispatchToProps
-)(ProfileDeletionWarningScreen);
+)(ProfileDeletionSummaryScreen);
